flask-api/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
gunicorn==21.2.0
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
moviepy==1.0.3
numpy==1.26.1
packaging==23.2
Pillow==10.0.1
proglog==0.1.10
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
urllib3==2.0.7
Werkzeug==3.0.1
